+++ packages/core/src/providers/index.ts
@@ -7,5 +7,8 @@ export { SessionProvider } from "./user/sessionProvider";
 export { WithdrawsProvider } from "./user/withdrawsProvider";
 export { SubscriptionProvider } from "./user/subscription";
 export { ConnectWalletProvider } from "./user/connectWalletProvider";
-export { DirectDepositProvider, DirectWithdrawProvider } from "./user/direct";
+export {
+  DirectDepositProvider,
+  DirectWithdrawProvider,
+} from "./user/sendAndReceive";
 export * from "./user/thea";
